requirements.txt
alpaca-trade-api>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
websockets<11,>=9.0
aiohttp>=3.8.5
gql>=3.4.1
asyncio>=3.4.3
requests>=2.31.0
pytest>=7.0.0  # For running the WebSocket test script
python-json-logger>=2.0.0  # For structured JSON logging
tenacity>=8.0.0  # For retry handling
aiodns>=3.1.1
# cchardet>=2.1.7  # Removed due to build issues
uvicorn>=0.24.0
fastapi>=0.104.1